% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Accident Map Visualization}
\usage{
fars_map_state(state.num, year)
}
\value{
The function returns a map of each US state (as identified by the state number in the 
function call) and the corresponding year. The map provides an outlined image of the state with
accidents plotted along their longitudinal and latitudinal coordinates
}
\description{
This function provides a map of the individual states and plots each accident on its latitudue and
longitude coordinates. The function requires both the state number and the year to function 
correctly.
}
\examples{
fars_map_state(53,2013)

}
