% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Monthly Accident Summary}
\usage{
fars_summarize_years(years)
}
\value{
The function returns a list of the total accidents for each individual month per year.
}
\description{
This function provides a summary of the total monthly accidents per year. The new list returned
by the function provides a sum of all accidents in all twelve months per year. The years can be
specified within the function call.
}
\examples{
fars_summarize_years(2013)
}
