% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create custom filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer respresenting the year to be used in the filename string}
}
\value{
This function returns a string with the filename customized with given year
}
\description{
This function creates a custom filename string from the combination of year given by the user (using \code{year}) and
the default part of the filename of the file with data from the
US National Highway Traffic Safety Administration's Fatality Analysis Reporting System.
This string can then be used to read data from the file with data from the given year.
The user input needs to be an integer otherwise the function will result in an error.
}
\examples{
make_filename(2013)
}
