% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read data from file into a dataframe.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string with the filename of the file to be read.}
}
\value{
This function returns a dataframe with the data read from the file.
}
\description{
This function reads data from a file given specified by the user and puts it in a dataframe.
If wrong filename is given then it will result in an error.
The function tbl_df() from dplyr is used.
}
